pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "forge-eval"
version = "0.1.0"
description = "Deterministic fail-closed Forge eval loop foundation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "jsonschema>=4.22.0",
  "PyYAML>=6.0.1",
]

[project.scripts]
forge-eval = "forge_eval.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-q"
pythonpath = ["src"]
markers = [
  "integration: integration tests",
]
